+++ src/main/java/com/TheWandererRaven/ravencoffee/customClasses/BrewEffect.java
@@ -1,16 +1,13 @@
 package com.TheWandererRaven.ravencoffee.customClasses;
 
 import net.minecraft.world.effect.MobEffect;
-import net.minecraft.world.effect.MobEffects;
-
-import java.util.List;
 
 public class BrewEffect {
     public MobEffect effect;
-    public double duration;
+    public int duration;
     public double chance;
     public boolean isForRemoving;
-    public BrewEffect(MobEffect _effect, double _duration, double _chance) {
+    public BrewEffect(MobEffect _effect, int _duration, double _chance) {
         this.effect = _effect;
         this.duration = _duration;
         this.chance = _chance;
@@ -18,7 +15,7 @@ public class BrewEffect {
     }
     public BrewEffect(MobEffect _effect) {
         this.effect = _effect;
-        this.duration = 0.0d;
+        this.duration = 0;
         this.chance = 0.0d;
         this.isForRemoving = true;
     }
